Development: change 'ReSharper enable' comments to 'ReSharper restore' (now working correctly) RIS: rewrite 'GetSize' methods (now return int and support any types (including reference structures) in Environment class RIS.Collections.Trees: add Trie and Trie<T> classes RIS.Extensions: add 'IndexOfAnyLinq' methods (old 'IndexOfAny' methods) in StringExtensions class & add 'IndexOfAnyTrie' methods in StringExtensions class & 'IndexOfAny' method now use 'IndexOfAnyTrie' internal in StringExtensions class & add length parameter for all 'IndexOfAny' methods in StringExtensions class & add 'GetAssemblyQualifiedName' extension method in TypeExtensions class RIS.Extensions.Entities: add null checks in RootIndex and LeafIndex classes RIS.Extensions.Frames: add TypeGetAssemblyQualifiedNameFrame class (for 'GetAssemblyQualifiedName' extension method for Type) RIS.Collections: add 'RIS.Collections.Extensions' namespace & add IKeyedEnumerable interface & IKeyedEnumerable<TKey, TValue> interface RIS.Collections.Extensions: add stack-like extensions methods ('Push', 'Pop', 'Peek', 'PeekRef', 'IsEmpty') in ChunkedCollectionExtensions class RIS.Collections.Chunked: remove redefinition 'SyncRoot' property in ISyncChunkedCollection interface RIS.Collections.Nestable: full rewrite and refactoring all code (for collections, serializers, helpers, etc.) (remove recursion calls, optimize code, improve perfomance, add new features, fix some bugs, change interfaces, add constructor with 'represent' parameter for collections, add new iterators for dictionaries, etc.) RIS.Collections.Trees: rename 'IKDPoinTComparer' interface to 'IKDPointComparer' & rename 'INeedsInitializationKDPoinTComparer' interface to 'IInitializableKDPointComparer' RIS.Settings: small refactoring in Setting class

+++ RIS.Collections/Chunked/Array/ChunkedArraySD.cs
@@ -126,7 +126,7 @@ namespace RIS.Collections.Chunked
 
         }
         public ChunkedArraySD(long length)
-            : this(length, Environment.GCLOHThresholdSize / Environment.GetSize<T>())
+            : this(length, Environment.GCLOHThresholdSize / (uint)Environment.GetSize<T>())
         {
 
         }
